pyproject.toml
[tool.poetry]
name = "expenses-tracker"
version = "0.1.0"
description = "Script to register expenses that comes via email to a google docs spreedsheet"
authors = ["Jairo Leyton <jali1993@hotmail.com>"]
readme = "README.md"
packages = [{include = "expenses_tracker"}]

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.71.0"
python-dotenv = "^0.21.1"
pytest = "^7.4.1"
beautifulsoup4 = "^4.12.2"
fastapi = "^0.110.0"
uvicorn = {version = "^0.27.1 ", extras = ["standard"]}
sqlmodel = "^0.0.16"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
